+++ src/app/adapters/defaultPlugins.js
@@ -1,8 +1,8 @@
 // @flow
 
 import type {StaticPluginAdapter} from "./pluginAdapter";
-import {StaticPluginAdapter as GitAdapter} from "../plugins/git/pluginAdapter";
-import {StaticPluginAdapter as GithubAdapter} from "../plugins/github/pluginAdapter";
+import {StaticPluginAdapter as GitAdapter} from "../../plugins/git/pluginAdapter";
+import {StaticPluginAdapter as GithubAdapter} from "../../plugins/github/pluginAdapter";
 
 export function defaultStaticAdapters(): $ReadOnlyArray<StaticPluginAdapter> {
   return [new GitAdapter(), new GithubAdapter()];
